feat- update the filepath of restructured components
